+++ CHANGES.md
@@ -28,6 +28,7 @@
     -   Fix `markdown-table-forward-cell` at last column issue [GH-522][]
     -   Fix GFM bold fontification with underscore issue [GH-525][]
     -   Fix wrong fontification words between strong markups [GH-534][]
+    -   Fix wrong italic fontification just after code block [GH-548][]
 
   [gh-290]: https://github.com/jrblevin/markdown-mode/issues/290
   [gh-311]: https://github.com/jrblevin/markdown-mode/issues/311
@@ -43,6 +44,7 @@
   [gh-532]: https://github.com/jrblevin/markdown-mode/issues/532
   [gh-534]: https://github.com/jrblevin/markdown-mode/issues/534
   [gh-536]: https://github.com/jrblevin/markdown-mode/issues/536
+  [gh-548]: https://github.com/jrblevin/markdown-mode/issues/548
 
 # Markdown Mode 2.4